Changes baseweight of DPI synapse to 7 pA.
There was a discrepancy in the baseweight variable between DPI and DPIstdp. Due to intense simulations
I changed the default DPI parameter to match the one of the DPIstdp synapse. Since the baseweight is only a
scaling factor nothing changes except the weight needs to be higher (approx. 7 times larger).