do not remove spike timing offset in dense representation in plotter2d anymore
so far, we subtracted the first spike time from all spike times before the conversion to the dense representation of the spike times. This saves memory, as there are no unnecessary zeros, but it is also unexpected. Now, the user has to remove the offset himself or use the plotrange.