Rename files and factorize code for tutorials.
I tested all files in theis commit to run. I documented all stps in the tutorial section of our documentation.
I renamed variables to follow pep8 and to be more easy to read.
Deleted poisson_test as it only contained a standard brian2 poission spike generator and should not be in teili, as this is a standard example of brian2.