Remove plot method.
As with the next master merge we will use the visualizer as designed by @rekrau, I removed
the plot mehtod for now. As soon as new visualizer is tested and incorporated in tutorials etc.
we will generate new plot mehtod for building blocks which use the underlying plot functions of the visualizer.