Update install instructions
After removin the `user_dir` option from the `setup.py` as it caused installation failure, I updated the corresponding instructions in the documentation. Further I added notes
to make the documentation easier to read. Secondly, I removed the `.py` ending in the instructions, how to generate models after installation, as it was infact wrong (syntax-wise)
Thanks @germain for spotting this.