Add note and update names in one code block.
Add attention note for weight value. As the baseweight is initilised to 7 pA at the moment the weight set to the synapse needs to be rather big.
In the future we should think about making the default basweight much greater, such that the users don't have to set such a large number.
One code block in the tutorial still had the old naming convention. I adjusted it accrodingly.