Removes whitespace and codegen target
A whitespace in front of a string when setting a header and using genn causes a seg fault.
Secondly, this commit removs the following line:
```python
prefs['codegen.cpp.extra_compile_args_gcc'].append('-std=c++14')
```
The c++14 standard set in the random_sampling function is not supported in genn.